fix: 修正UserRole实体时间字段命名错误和TaskTemplate枚举字段缺失

+++ app/src/main/java/online/longlian/app/service/resource/FileStorageService.java
@@ -53,7 +53,7 @@ public class FileStorageService {
                 .bizType(createFileReqDTO.getBizType())
                 .bizId(createFileReqDTO.getBizId())
                 .processStatus(FileProcessStatus.UN_PROCESS)
-                .isReferenced(1)
+                .isReferenced((byte) 1)
                 .creatorId(ThreadLocalUtil.getUserBO().getId())
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
